tests: Move `PUT /crates/:name/owners` tests to `krate::owners` module
